Move functionality function to utils folder
