+++ Source/Utils/functionality.js
@@ -1,3 +1,5 @@
+const moment = require("moment");
+
 /**
  * This function allows us to manage the pagination.
  *
@@ -52,8 +54,74 @@ const escapeRegex = (string) => {
     return string.replace(/[-[\]{}()*+?.,\\^$|#\s]/g, "\\$&");
 };
 
+/**
+ * Create_structure function : initialization of the response object structure.
+ *
+ * @param {string} resource_type - Resource type : domain entities information.
+ * @param {object} resource - Response body object.
+ * @param {object} links - Link resource information : HATEOAS.
+ * @returns {object} Structure - Final response body object structure.
+ */
+function create_structure(resource_type, resource, links) {
+    const structure = {};
+    structure._resource_type = (resource_type !== '' ? resource_type : 'Resource_Error');
+    structure._resource = resource;
+    structure._links = links;
+    structure._etag = moment().format('MMMM Do YYYY, h:mm:ss a');
+    return structure;
+}
+
+/**
+ * Message_error function.
+ *
+ * @param {object} options - Errors message options fields.
+ * @returns {object} Message_error_object - Error standard object structure.
+ */
+function message_error(options) {
+    const message_error_object = {};
+    message_error_object._api_status_code = options.api_status_code || 4000;
+    message_error_object._api_status_message = options.api_status_message || 'Default Message Error';
+    message_error_object._api_status_id = options.request_id || '00000-00000-00000-00000-00000'; // TODO apply the request_id logic, you can use the same correlation id with app insight to propagate the context information across all your services.
+    message_error_object._details = options.details || [];
+    console.error(`${message_error_object._api_status_message}`);
+    return message_error_object;
+}
+
+/**
+ * creat_hateoas_structure function.
+ * TODO : you have to configure HATEOAS based on your needs. You find below some examples and explanation
+ * This function implement the RFC 5988 for building links that define the relationships between our resources.
+ * Each resource in RFC 5988 contains the following properties: { href: Target URI, rel: Link relation type, type: Attributes for target IRI }
+ * Use case :
+ * We can configure this section to determine what actions a user has access to.
+ * Therefore, the client does not need to know anything about roles or states of entities.
+ * All the actions on the screen are enabled/disabled/visible based on the presence of links.
+ * @param {object} request - Express request object.
+ * @returns {object} hateoas_structure - The hateoas structure RFC 5988 format.
+ */
+function creat_hateoas_structure(request, collection = false) {
+    const hateoas_object = {};
+    const links_tab = [];
+    const self = {
+        href: `${request.protocol + '://' + request.get('host') + request.originalUrl}`,
+        rel: `self`,
+        type: `${request.method}`,
+    };
+    links_tab.push(self);
+    if(collection) {
+        hateoas_object.web_pages = parse_url_page(self.href, request.query.offset, request.query.limit, request.query.count);
+    }
+    hateoas_object.web_links = links_tab;
+    // you can add some hateoas logic here
+
+    return hateoas_object;
+}
+
 module.exports = {
     parse_url_page,
     extend,
     escapeRegex,
+    message_error,
+    create_structure,
+    creat_hateoas_structure
 };
